requirements.txt
psycopg==3.1.7
apache-airflow==2.5.0
loguru==0.6.0
